+++ src/BrandCard/BrandCard.tsx
@@ -1,9 +1,9 @@
 import React, { FC, ReactNode } from 'react'
 import styled from 'styled-components'
-import { MarginProps } from '../utils/space'
-import { Color, theme } from '../theme'
 import { Box } from '../Box'
 import { Text } from '../Text'
+import { Color, theme } from '../theme'
+import { MarginProps } from '../utils/space'
 
 export type BrandCardProps = {
   /** generic card title */
@@ -18,7 +18,7 @@ export type BrandCardProps = {
   visualHeight?: string
   /** render image at bottom of card */
   visualBottom?: boolean
-  bgColour: Color
+  bgColor: Color
   /** card tag */
   tag?: ReactNode
   /** primary button */
@@ -39,25 +39,24 @@ export const BrandCard: FC<BrandCardProps> = ({
   tag,
   buttonAction,
   fallbackStyle = false,
-  bgColour = 'matcha',
+  bgColor = 'matcha',
   maxWidth = '',
   ...otherProps
 }) => {
   return (
     <Container
       maxWidth={maxWidth}
-      visual={visual}
-      fallbackStyle={fallbackStyle}
-      bgColor={bgColour}
+      $fallbackStyle={fallbackStyle}
+      $bgColor={bgColor}
       {...otherProps}
     >
       {tag && <TagWrapper>{tag}</TagWrapper>}
       {visual && !visualBottom && (
-        <VisualWrapper visualHeight={visualHeight}>
+        <VisualWrapper $visualHeight={visualHeight}>
           <Visual
-            visualUrl={visual}
-            visualHeight={visualHeight}
-            alignVisual={alignVisual}
+            $visualUrl={visual}
+            $visualHeight={visualHeight}
+            $alignVisual={alignVisual}
           />
         </VisualWrapper>
       )}
@@ -79,11 +78,11 @@ export const BrandCard: FC<BrandCardProps> = ({
       </Box>
       {buttonAction && <Box mt="16px">{buttonAction}</Box>}
       {visual && visualBottom && (
-        <VisualWrapper visualHeight={visualHeight} mt="16px">
+        <VisualWrapper $visualHeight={visualHeight} mt="16px">
           <Visual
-            visualUrl={visual}
-            visualHeight={visualHeight}
-            alignVisual={alignVisual}
+            $visualUrl={visual}
+            $visualHeight={visualHeight}
+            $alignVisual={alignVisual}
           />
         </VisualWrapper>
       )}
@@ -92,12 +91,11 @@ export const BrandCard: FC<BrandCardProps> = ({
 }
 
 const Container = styled(Box)<{
-  visual?: string
-  fallbackStyle?: boolean
-  bgColor: Color
+  $fallbackStyle?: boolean
+  $bgColor: Color
 }>`
-  background: ${({ fallbackStyle, bgColor }) =>
-    fallbackStyle ? theme.colors.cream : theme.colors[bgColor]};
+  background: ${({ $fallbackStyle, $bgColor }) =>
+    $fallbackStyle ? theme.colors.cream : theme.colors[$bgColor]};
   box-sizing: border-box;
   border-radius: 16px 16px 0px 0px;
 
@@ -112,24 +110,24 @@ const TagWrapper = styled(Box)`
   right: 16px;
 `
 
-const VisualWrapper = styled(Box)<{ visualHeight: string }>`
+const VisualWrapper = styled(Box)<{ $visualHeight: string }>`
   width: 100%;
 
-  ${({ visualHeight }) => visualHeight && `height: ${visualHeight};`}
+  ${({ $visualHeight }) => $visualHeight && `height: ${$visualHeight};`}
 `
 
 const Visual = styled(Box)<{
-  visualUrl: string
-  visualHeight: string
-  alignVisual?: string
+  $visualUrl: string
+  $visualHeight: string
+  $alignVisual?: string
 }>`
   width: 100%;
-  background-image: url('${(p) => p.visualUrl}');
+  background-image: url('${(p) => p.$visualUrl}');
   background-size: contain;
   background-repeat: no-repeat;
-  background-position: ${({ alignVisual }) =>
-    alignVisual ? alignVisual : 'left'};
+  background-position: ${({ $alignVisual }) =>
+    $alignVisual ? $alignVisual : 'left'};
 
-  ${({ visualHeight }) =>
-    visualHeight ? `height: ${visualHeight};` : 'padding-top: 100%;'}
+  ${({ $visualHeight }) =>
+    $visualHeight ? `height: ${$visualHeight};` : 'padding-top: 100%;'}
 `