+++ src/Button/Button.tsx
@@ -1,6 +1,7 @@
 import React, { ButtonHTMLAttributes, FC, ReactNode, forwardRef } from 'react'
 import styled, { css } from 'styled-components'
 
+import { TransientProps } from 'utils/utilTypes'
 import { Box } from '../Box'
 import { Icon as IconComponent } from '../Icon'
 import { Icons } from '../Icon/iconsList'
@@ -58,20 +59,19 @@ export const Button: FC<ButtonProps> = forwardRef<
 
   return (
     <Container
-      as="button"
+      forwardedAs="button"
       id={id}
       className={className}
       disabled={disabled || loading}
       onClick={handleClick}
       $loading={loading}
-      primary={primary}
-      secondary={secondary}
-      fallbackStyle={fallbackStyle}
-      textBtn={textBtn}
-      smallButton={smallButton}
-      icon={icon}
-      trailingIcon={trailingIcon}
-      forcedWidth={forcedWidth}
+      $primary={primary}
+      $secondary={secondary}
+      $fallbackStyle={fallbackStyle}
+      $textBtn={textBtn}
+      $smallButton={smallButton}
+      $trailingIcon={trailingIcon}
+      $forcedWidth={forcedWidth}
       {...(form ? { form } : {})}
       type={type}
       {...otherProps}
@@ -82,10 +82,10 @@ export const Button: FC<ButtonProps> = forwardRef<
           <Loader color={'liquorice'} height="16" />
         </LoaderContainer>
       )}
-      <ContentContainer icon={icon} $loading={loading}>
+      <ContentContainer $icon={icon} $loading={loading}>
         {!trailingIcon && icon && (
           <IconContainer
-            trailingIcon={trailingIcon}
+            $trailingIcon={trailingIcon}
             render={icon}
             size={smallButton ? 16 : 24}
             color={'liquorice'}
@@ -96,7 +96,7 @@ export const Button: FC<ButtonProps> = forwardRef<
         </ChildrenContainer>
         {trailingIcon && icon && textBtn && (
           <IconContainer
-            trailingIcon={trailingIcon}
+            $trailingIcon={trailingIcon}
             render={icon}
             size={smallButton ? 16 : 24}
             color={'liquorice'}
@@ -110,33 +110,34 @@ export const Button: FC<ButtonProps> = forwardRef<
 
 Button.displayName = 'Button'
 
-type IButton = Required<
-  Pick<
-    ButtonProps,
-    | 'disabled'
-    | 'primary'
-    | 'secondary'
-    | 'forcedWidth'
-    | 'fallbackStyle'
-    | 'textBtn'
-    | 'trailingIcon'
-    | 'smallButton'
+type IButton = TransientProps<
+  Required<
+    Pick<
+      ButtonProps,
+      | 'primary'
+      | 'secondary'
+      | 'forcedWidth'
+      | 'fallbackStyle'
+      | 'textBtn'
+      | 'trailingIcon'
+      | 'smallButton'
+    >
   >
 > & {
   $loading: NonNullable<ButtonProps['loading']>
-  icon?: Icons
+  disabled: boolean
 }
 
 const Container = styled(Box)<IButton>(
   ({
     disabled,
     $loading,
-    primary,
-    secondary,
-    forcedWidth,
-    fallbackStyle,
-    textBtn,
-    smallButton,
+    $primary,
+    $secondary,
+    $forcedWidth,
+    $fallbackStyle,
+    $textBtn,
+    $smallButton,
   }) => css`
     position: relative;
     background-color: ${theme.colors.marshmallowPink};
@@ -150,11 +151,11 @@ const Container = styled(Box)<IButton>(
     line-height: 100%;
     font-size: 16px;
     opacity: ${disabled ? '0.5' : '1'};
-    width: ${forcedWidth ? forcedWidth : 'auto'};
+    width: ${$forcedWidth ? $forcedWidth : 'auto'};
 
     ${focusOutlineStyle}
 
-    ${primary &&
+    ${$primary &&
     css`
       color: ${theme.colors.liquorice};
 
@@ -165,7 +166,7 @@ const Container = styled(Box)<IButton>(
         background-color: ${theme.colors.lollipop};
       }
     `}
-    ${secondary &&
+    ${$secondary &&
     css`
       background-color: ${theme.colors.oatmeal};
 
@@ -176,7 +177,7 @@ const Container = styled(Box)<IButton>(
         background-color: ${theme.colors.custard};
       }
     `}
-  ${fallbackStyle &&
+  ${$fallbackStyle &&
     css`
       background-color: ${theme.colors.cream};
 
@@ -187,7 +188,7 @@ const Container = styled(Box)<IButton>(
         background-color: ${theme.colors.mascarpone};
       }
     `}
-  ${smallButton &&
+  ${$smallButton &&
     css`
       padding: 0 10px;
       min-width: 54px;
@@ -201,7 +202,7 @@ const Container = styled(Box)<IButton>(
         margin: 0 5px 0 0;
       }
     `}
-  ${textBtn &&
+  ${$textBtn &&
     css`
       background-color: transparent;
       padding: 0;
@@ -230,18 +231,18 @@ const LoaderContainer = styled.div`
   justify-content: center;
 `
 
-const ContentContainer = styled.div<
-  Pick<ButtonProps, 'icon'> & { $loading: boolean }
->`
+const ContentContainer = styled.div<{ $loading: boolean; $icon?: Icons }>`
   display: flex;
   align-items: center;
-  justify-content: ${({ icon }) => (icon ? 'space-evenly' : 'center')};
+  justify-content: ${({ $icon }) => ($icon ? 'space-evenly' : 'center')};
   opacity: ${({ $loading }) => ($loading ? '0' : '1')};
 `
 
-const IconContainer = styled(IconComponent)<Pick<ButtonProps, 'trailingIcon'>>(
-  ({ trailingIcon }) => css`
-    margin: ${trailingIcon ? '0 0 0 10px' : '0 10px 0 0'};
+const IconContainer = styled(IconComponent)<
+  TransientProps<Pick<ButtonProps, 'trailingIcon'>>
+>(
+  ({ $trailingIcon }) => css`
+    margin: ${$trailingIcon ? '0 0 0 10px' : '0 10px 0 0'};
   `,
 )