+++ src/Table/components/TableHeader.tsx
@@ -17,12 +17,12 @@ export const TableHeader = <T extends object>({
       {columns.map((column, columnIndex) => (
         <StyledHeaderCell
           key={columnIndex}
-          fixedHeader={fixedHeader}
-          headerHeight={headerHeight}
-          minWidth={column.minWidth}
-          maxWidth={column.maxWidth}
-          headerColor={headerColor}
-          columnPadding={columnPadding}
+          $fixedHeader={fixedHeader}
+          $headerHeight={headerHeight}
+          $minWidth={column.minWidth}
+          $maxWidth={column.maxWidth}
+          $headerColor={headerColor}
+          $columnPadding={columnPadding}
           $hasKeyline={hasKeyline}
         >
           {column.name}
@@ -30,11 +30,11 @@ export const TableHeader = <T extends object>({
       ))}
       {(expandable || rowActions) && (
         <StyledHeaderCell
-          fixedHeader={fixedHeader}
-          stickyCell={true}
-          minWidth={rowActions?.minWidth}
-          headerColor={headerColor}
-          columnPadding={columnPadding}
+          $fixedHeader={fixedHeader}
+          $stickyCell={true}
+          $minWidth={rowActions?.minWidth}
+          $headerColor={headerColor}
+          $columnPadding={columnPadding}
           $hasKeyline={hasKeyline}
         >
           Actions